added K8S services for proxy tunnel mode
Additionally, it is possible to distribute several ZitiServices and
their listen ports to different Kubernetes services. Use cases are those
where you want, for example, one zitiService to be published on one port
that should be on a ClusterIP service, when another should be on a
LoadBalancer or NodePort service.